Feat: api 명세 추가

+++ src/main/java/com/example/savingsalt/community/comment/controller/CommentController.java
@@ -1,10 +1,12 @@
 package com.example.savingsalt.community.comment.controller;
 
-import com.example.savingsalt.community.board.exception.BoardException.UnauthorizedCreateException;
 import com.example.savingsalt.community.comment.domain.dto.CommentReqDto;
 import com.example.savingsalt.community.comment.domain.dto.CommentResDto;
 import com.example.savingsalt.community.comment.service.CommentService;
+import com.example.savingsalt.global.UnauthorizedException;
 import com.example.savingsalt.member.domain.MemberEntity;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -16,6 +18,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+@Tag(name = "Comment", description = "댓글 API")
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api")
@@ -23,40 +26,46 @@ public class CommentController {
 
     private final CommentService commentService;
 
+    @Operation(summary = "댓글 작성", description = "로그인된 사용자가 새로운 댓글을 작성합니다.")
     @PostMapping("/comments")
     public ResponseEntity<CommentResDto> createComment(@RequestBody CommentReqDto requestDto,
-        @AuthenticationPrincipal
-        MemberEntity member) {
+        @AuthenticationPrincipal MemberEntity member) {
+
         if (member == null) {
-            throw new UnauthorizedCreateException();
+            throw new UnauthorizedException("로그인이 필요합니다.");
         }
 
-        CommentReqDto updateReqDto = CommentReqDto.builder()
-            .content(requestDto.getContent())
-            .boardId(requestDto.getBoardId())
-            .memberId(member.getId())
-            .build();
-
-        CommentResDto responseDto = commentService.createComment(updateReqDto, member);
+        CommentResDto responseDto = commentService.createComment(requestDto, member);
 
         return ResponseEntity.ok(responseDto);
     }
 
+//    @GetMapping("/comments/{commentId}")
+//    public List<CommentResDto> getAllComments(@PathVariable Long commentId) {
+//        return commentService.getCommentsByBoardId(commentId);
+//    }
+
+    @Operation(summary = "댓글 수정", description = "댓글 ID를 통해 댓글의 내용을 수정합니다.")
     @PatchMapping("/comments/{commentId}")
     public ResponseEntity<String> updateComment(@PathVariable Long commentId,
         @RequestBody CommentReqDto requestDto, @AuthenticationPrincipal MemberEntity member) {
 
         if (member == null) {
-            throw new UnauthorizedCreateException();
+            throw new UnauthorizedException("로그인이 필요합니다.");
         }
 
         return commentService.updateComment(commentId, requestDto, member);
     }
 
+    @Operation(summary = "댓글 삭제", description = "댓글 ID를 통해 댓글을 삭제합니다.")
     @DeleteMapping("/comments/{commentId}")
     public ResponseEntity<String> deleteComment(@PathVariable Long commentId,
         @AuthenticationPrincipal MemberEntity member) {
+
+        if (member == null) {
+            throw new UnauthorizedException("로그인이 필요합니다.");
+        }
+
         return commentService.deleteComment(commentId, member);
     }
-
 }
